added precision setting for deployment and changed 32-mixed to 32-true since 32-mixed doesnt exist

+++ src/stamp/modeling/train.py
@@ -215,7 +215,7 @@ def train_model_(
         )
         precision="16-mixed"
     else:
-        precision="32-mixed"
+        precision="32-true"
         torch.set_float32_matmul_precision("high")
     
     trainer = lightning.Trainer(
